Move `SimpleContextKey` tests to separate file
