+++ src/context-registry.spec.ts
@@ -1,8 +1,7 @@
 import { AIterable } from 'a-iterable';
-import { ContextKeyError } from './context-key-error';
 import { ContextRegistry } from './context-registry';
 import { ContextValues } from './context-values';
-import { MultiContextKey, SingleContextKey } from './simple-context-key';
+import { SingleContextKey } from './simple-context-key';
 import Mock = jest.Mock;
 
 describe('ContextRegistry', () => {
@@ -47,217 +46,6 @@ describe('ContextRegistry', () => {
 
       expect([...values.get(key.seedKey)]).toEqual([value]);
       expect(values.get(key)).toBe(value);
-    });
-  });
-
-  describe('Single value', () => {
-    it('provides value', () => {
-
-      const value = 'test value';
-
-      mockProvider.mockReturnValue(value);
-
-      expect(values.get(key)).toBe(value);
-    });
-    it('selects the last value if more than one provided', () => {
-
-      const value1 = 'value1';
-      const value2 = 'value2';
-
-      registry.provide({ a: key, is: value1 });
-      registry.provide({ a: key, is: value2 });
-
-      expect(values.get(key)).toBe(value2);
-    });
-    it('removes the value specifier', () => {
-
-      const value1 = 'value1';
-      const value2 = 'value2';
-
-      registry.provide({ a: key, is: value1 });
-      registry.provide({ a: key, is: value2 })();
-
-      expect(values.get(key)).toBe(value1);
-    });
-    it('retains the constructed value when specifier removed', () => {
-
-      const value1 = 'value1';
-      const value2 = 'value2';
-
-      registry.provide({ a: key, is: value1 });
-
-      const remove = registry.provide({ a: key, is: value2 });
-
-      expect(values.get(key)).toBe(value2);
-
-      remove();
-      remove();
-      expect(values.get(key)).toBe(value2);
-    });
-    it('throws if there is neither default nor fallback value', () => {
-      expect(() => values.get(new SingleContextKey(key.name))).toThrowError(ContextKeyError);
-      expect(() => values.get(new SingleContextKey(key.name), {})).toThrowError(ContextKeyError);
-    });
-    it('provides fallback value if there is no provider', () => {
-      expect(values.get(new SingleContextKey<string>(key.name), { or: 'fallback' })).toBe('fallback');
-    });
-    it('provides default value if provider did not provide any value', () => {
-
-      const defaultValue = 'default';
-      const keyWithDefaults = new SingleContextKey(key.name, { byDefault: () => defaultValue });
-
-      registry.provide({ a: keyWithDefaults, is: null });
-
-      expect(values.get(keyWithDefaults)).toBe(defaultValue);
-    });
-    it('provides default value if there is no provider', () => {
-      expect(values.get(new SingleContextKey<string>(key.name, { byDefault: () => 'default' }))).toBe('default');
-    });
-    it('prefers fallback value over default one', () => {
-      expect(values.get(new SingleContextKey<string>(key.name, { byDefault: () => 'default' }), { or: 'fallback' }))
-          .toBe('fallback');
-    });
-    it('prefers default value if fallback one is absent', () => {
-      expect(values.get(new SingleContextKey<string>(key.name, { byDefault: () => 'default' }), {}))
-          .toBe('default');
-    });
-    it('prefers `null` fallback value over key one', () => {
-      expect(values.get(new SingleContextKey<string>(key.name, { byDefault: () => 'default' }), { or: null }))
-          .toBeNull();
-    });
-    it('prefers `undefined` fallback value over key one', () => {
-      expect(values.get(new SingleContextKey<string>(key.name, { byDefault: () => 'default' }), { or: undefined }))
-          .toBeUndefined();
-    });
-    it('caches the value', () => {
-
-      const value = 'value';
-
-      mockProvider.mockReturnValue(value);
-
-      expect(values.get(key)).toBe(value);
-      expect(values.get(key)).toBe(value);
-
-      expect(mockProvider).toHaveBeenCalledTimes(1);
-    });
-    it('caches default value', () => {
-
-      const value = 'default value';
-      const defaultProviderSpy = jest.fn(() => value);
-      const keyWithDefault = new SingleContextKey('key-with-default', { byDefault: defaultProviderSpy });
-
-      expect(values.get(keyWithDefault)).toBe(value);
-      expect(values.get(keyWithDefault)).toBe(value);
-      expect(defaultProviderSpy).toHaveBeenCalledTimes(1);
-    });
-    it('does not cache fallback value', () => {
-
-      const value1 = 'value1';
-      const value2 = 'value2';
-
-      expect(values.get(key, { or: value1 })).toBe(value1);
-      expect(values.get(key, { or: value2 })).toBe(value2);
-    });
-    it('rebuilds the value in another context', () => {
-
-      const value1 = 'value1';
-      const value2 = 'value2';
-
-      mockProvider.mockReturnValue(value1);
-      expect(values.get(key)).toBe(value1);
-
-      const values2 = registry.newValues();
-
-      mockProvider.mockReturnValue(value2);
-      expect(values.get(key)).toBe(value1);
-      expect(values2.get(key)).toBe(value2);
-
-      expect(mockProvider).toHaveBeenCalledTimes(2);
-    });
-  });
-
-  describe('Multi-value', () => {
-
-    let multiKey: MultiContextKey<string>;
-
-    beforeEach(() => {
-      multiKey = new MultiContextKey('values');
-    });
-
-    it('is associated with empty array by default', () => {
-      expect(values.get(multiKey)).toEqual([]);
-    });
-    it('is associated with empty array if providers did not return any values', () => {
-      registry.provide({ a: multiKey, is: null });
-      registry.provide({ a: multiKey, is: undefined });
-
-      expect(values.get(multiKey)).toEqual([]);
-    });
-    it('is associated with default value if there is no provider', () => {
-
-      const defaultValue = ['default'];
-      const keyWithDefaults = new MultiContextKey('key', { byDefault: () => defaultValue });
-
-      expect(values.get(keyWithDefaults)).toEqual(defaultValue);
-    });
-    it('is associated with default value if providers did not return any values', () => {
-
-      const defaultValue = ['default'];
-      const keyWithDefaults = new MultiContextKey('key', { byDefault: () => defaultValue });
-
-      registry.provide({ a: keyWithDefaults, is: null });
-      registry.provide({ a: keyWithDefaults, is: undefined });
-
-      expect(values.get(keyWithDefaults)).toEqual(defaultValue);
-    });
-    it('is associated with provided values array', () => {
-      registry.provide({ a: multiKey, is: 'a' });
-      registry.provide({ a: multiKey, is: undefined });
-      registry.provide({ a: multiKey, is: 'c' });
-
-      expect(values.get(multiKey)).toEqual(['a', 'c']);
-    });
-    it('is associated with value', () => {
-      registry.provide({ a: multiKey, is: 'value' });
-
-      expect(values.get(multiKey)).toEqual(['value']);
-    });
-    it('throws if there is no default value', () => {
-      expect(() => values.get(new MultiContextKey(multiKey.name, { byDefault: () => null }))).toThrowError();
-    });
-    it('is associated with empty array by default', () => {
-      expect(values.get(new MultiContextKey(multiKey.name))).toEqual([]);
-    });
-    it('is associated with default value is there is no value', () => {
-      expect(values.get(new MultiContextKey<string>(multiKey.name), { or: ['default'] }))
-          .toEqual(['default']);
-    });
-    it('is associated with key default value is there is no value', () => {
-      expect(values.get(new MultiContextKey<string>(multiKey.name, { byDefault: () => ['default'] })))
-          .toEqual(['default']);
-    });
-    it('prefers fallback value over default one', () => {
-      expect(values.get(
-          new MultiContextKey<string>(
-              multiKey.name,
-              { byDefault: () => ['key', 'default'] }
-          ),
-          { or: ['explicit', 'default'] }
-      )).toEqual(['explicit', 'default']);
-    });
-    it('prefers `null` fallback value over default one', () => {
-      expect(values.get(
-          new MultiContextKey<string>(multiKey.name, { byDefault: () => ['key', 'default'] }),
-          { or: null }
-      )).toBeNull();
-    });
-    it('prefers `undefined` fallback value over default one', () => {
-      expect(values.get(new MultiContextKey<string>(
-          multiKey.name,
-          { byDefault: () => ['key', 'default'] }
-          ),
-          { or: undefined }
-      )).toBeUndefined();
     });
   });
